+++ src/problems/rrnrp/RecoverableRobustNextReleaseProblem.java
@@ -16,7 +16,7 @@ public class RecoverableRobustNextReleaseProblem{
 
    int numberOfRequirements;
    double[] requirementsValues;
-   int[] requirementsCosts;
+   double[] requirementsCosts;
    double[] requirementsDeviances;
    double budget;
    int gamma;
